Image standard...for now
For landscape images, 600px width looks best with our 600px content column. My preference would be a 300px width with text wrap, then the 300px links to the 1024px version. This is a quick and dirty fix for landscapes, but will be more problematic with portraits. Will write this up in the wiki.